Fizz Buzz
What is "Fizz Buzz"?
Fizz Buzz is a simple program that prints numbers from 1 to a given number (commonly 100). However, there are a few rules:

If a number is divisible by 3, the program prints "Fizz" instead of the number.
If a number is divisible by 5, the program prints "Buzz" instead of the number.
If a number is divisible by both 3 and 5, the program prints "FizzBuzz".